+++ Sanctuary.Census/Abstractions/Database/ICollectionsContext.cs
@@ -157,6 +157,14 @@ public interface ICollectionsContext
     Task UpsertMapRegionsAsync(IEnumerable<MapRegion> collection, CancellationToken ct = default);
 
     /// <summary>
+    /// Upserts the <see cref="OutfitWarRegistration"/> collection.
+    /// </summary>
+    /// <param name="collection">The collection.</param>
+    /// <param name="ct">A <see cref="CancellationToken"/> that can be used to stop the operation.</param>
+    /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
+    Task UpsertOutfitWarRegistrationsAsync(IEnumerable<OutfitWarRegistration> collection, CancellationToken ct = default);
+
+    /// <summary>
     /// Upserts the <see cref="PlayerStateGroup2"/> collection.
     /// </summary>
     /// <param name="collection">The collection.</param>