+++ Sanctuary.Census/CollectionBuilders/OutfitWarRegistratationCollectionBuilder.cs
@@ -1,0 +1,57 @@
+﻿using Sanctuary.Census.Abstractions.CollectionBuilders;
+using Sanctuary.Census.Abstractions.Database;
+using Sanctuary.Census.Exceptions;
+using Sanctuary.Census.Models.Collections;
+using Sanctuary.Census.ServerData.Internal.Abstractions.Services;
+using Sanctuary.Zone.Packets.OutfitWars;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Sanctuary.Census.CollectionBuilders;
+
+/// <summary>
+/// Builds the <see cref="OutfitWarRegistration"/> collection.
+/// </summary>
+public class OutfitWarRegistrationCollectionBuilder : ICollectionBuilder
+{
+    private readonly IServerDataCacheService _serverDataCache;
+
+    /// <summary>
+    /// Initializes a new instance of the <see cref="OutfitWarRegistrationCollectionBuilder"/> class.
+    /// </summary>
+    /// <param name="serverDataCache">The server data cache.</param>
+    public OutfitWarRegistrationCollectionBuilder
+    (
+        IServerDataCacheService serverDataCache
+    )
+    {
+        _serverDataCache = serverDataCache;
+    }
+
+    /// <inheritdoc />
+    public async Task BuildAsync
+    (
+        ICollectionsContext dbContext,
+        CancellationToken ct
+    )
+    {
+        if (_serverDataCache.RegisteredOutfits is null)
+            throw new MissingCacheDataException(typeof(RegisteredOutfits));
+
+        Dictionary<ulong, OutfitWarRegistration> builtOutfits = new();
+        foreach (RegisteredOutfit outfit in _serverDataCache.RegisteredOutfits.Outfits)
+        {
+            OutfitWarRegistration built = new
+            (
+                outfit.OutfitID,
+                (uint)outfit.FactionID,
+                outfit.RegistrationOrder,
+                outfit.MemberSignupCount
+            );
+            builtOutfits.Add(built.OutfitID, built);
+        }
+
+        await dbContext.UpsertOutfitWarRegistrationsAsync(builtOutfits.Values, ct);
+    }
+}