+++ Sanctuary.Census/Database/CollectionsContext.cs
@@ -77,6 +77,7 @@ public class CollectionsContext : ICollectionsContext
         await CreateUniqueKeyIndex<Loadout>(x => x.LoadoutID, ct).ConfigureAwait(false);
         await CreateNonUniqueKeyIndexes<LoadoutSlot>(ct, x => x.LoadoutID, x => x.SlotID).ConfigureAwait(false);
         await CreateUniqueKeyIndex<MapRegion>(x => x.MapRegionId, ct).ConfigureAwait(false);
+        await CreateUniqueKeyIndex<OutfitWarRegistration>(x => x.OutfitID, ct).ConfigureAwait(false);
         await CreateNonUniqueKeyIndexes<PlayerStateGroup2>(ct, x => x.PlayerStateGroupId, x => x.PlayerStateId).ConfigureAwait(false);
         await CreateUniqueKeyIndex<Profile>(x => x.ProfileId, ct).ConfigureAwait(false);
         await CreateUniqueKeyIndex<Projectile>(x => x.ProjectileId, ct).ConfigureAwait(false);
@@ -318,6 +319,16 @@ public class CollectionsContext : ICollectionsContext
         ).ConfigureAwait(false);
 
     /// <inheritdoc />
+    public async Task UpsertOutfitWarRegistrationsAsync(IEnumerable<OutfitWarRegistration> collection, CancellationToken ct = default)
+        => await UpsertCollectionAsync
+        (
+            collection,
+            e => x => x.OutfitID == e.OutfitID,
+            e => Builders<OutfitWarRegistration>.Filter.Eq(x => x.OutfitID, e.OutfitID),
+            ct
+        ).ConfigureAwait(false);
+
+    /// <inheritdoc />
     public async Task UpsertPlayerStateGroup2Async(IEnumerable<PlayerStateGroup2> collection, CancellationToken ct = default)
         => await UpsertCollectionAsync
         (
@@ -505,6 +516,7 @@ public class CollectionsContext : ICollectionsContext
         BsonClassMap.RegisterClassMap<Loadout>(MongoContext.AutoClassMap);
         BsonClassMap.RegisterClassMap<LoadoutSlot>(MongoContext.AutoClassMap);
         BsonClassMap.RegisterClassMap<MapRegion>(MongoContext.AutoClassMap);
+        BsonClassMap.RegisterClassMap<OutfitWarRegistration>(MongoContext.AutoClassMap);
         BsonClassMap.RegisterClassMap<PlayerStateGroup2>(MongoContext.AutoClassMap);
         BsonClassMap.RegisterClassMap<Profile>(MongoContext.AutoClassMap);
         BsonClassMap.RegisterClassMap<Projectile>(MongoContext.AutoClassMap);